Used a default resolution for all control variables to improve accuracy

+++ lib/Xpilot/src/ModeStaticDef.cpp
@@ -5,23 +5,23 @@
  * We define them here instead
  */
 
-int16_t Mode::input_rpy[3]{0, 0, 0};
-int16_t Mode::output_rpy[3]{0, 0, 0};
+int32_t Mode::input_rpy[3]{0, 0, 0};
+int32_t Mode::output_rpy[3]{0, 0, 0};
 
-float Mode::imu_rpy[3]{0.f, 0.f, 0.f};
-float Mode::imu_g[3]{0.f, 0.f, 0.f};
+int32_t Mode::imu_rpy[3]{0, 0, 0};
+int32_t Mode::imu_g[3]{0, 0, 0};
 
 int16_t Mode::SRVout[Actuators::Channel::CHANNEL_COUNT]{0, 0, 0, 0};
 
-PIDF<int16_t> Mode::rollPIDF;
-PIDF<int16_t> Mode::pitchPIDF;
-PIDF<int16_t> Mode::yawPIDF;
+PIDF<int32_t> Mode::rollPIDF;
+PIDF<int32_t> Mode::pitchPIDF;
+PIDF<int32_t> Mode::yawPIDF;
 
 AirplaneMixer Mode::airplaneMixer;
 
-SlewRateLimiter<int16_t> Mode::rollSlew;
-SlewRateLimiter<int16_t> Mode::pitchSlew;
-SlewRateLimiter<int16_t> Mode::yawSlew;
+SlewRateLimiter<int32_t> Mode::rollSlew;
+SlewRateLimiter<int32_t> Mode::pitchSlew;
+SlewRateLimiter<int32_t> Mode::yawSlew;
 
 #if defined(USE_FLAPERONS)
 int16_t Mode::flaperonOut = 0;
